get install and linking cleaned up

+++ link_homedir.sh
@@ -10,7 +10,12 @@ DIR=$(set_script_dir)
 cd ~
 
 # Backup existing files (killing existing backups first)
-rm -rf $DIR/backups
+echo "Destroying old backups..."
+if [ -d $DIR/backups ]; then
+  rm -rf $DIR/backups
+fi
+
+echo "Making new backups..."
 mkdir -p $DIR/backups
 
 mv -f ~/.bash_profile $DIR/backups/
@@ -19,7 +24,27 @@ mv -f ~/.vimrc $DIR/backups/
 mv -f ~/.gitconfig $DIR/backups/
 
 # Set up symbolic links
+echo "Setting up symbolic links..."
 ln -s $DIR/.bash_profile
 ln -s $DIR/.vim
 ln -s $DIR/.vimrc
 ln -s $DIR/.gitconfig
+
+# Create backup and temp folders for vim if they don't exist already
+echo "Creating _backup and _temp folders for vim..."
+if [ ! -d $DIR/.vim/_backup ]; then
+  echo "Creating $DIR/.vim/_backup..."
+  mkdir $DIR/.vim/_backup
+fi
+
+if [ ! -d $DIR/.vim/_temp ]; then
+  echo "Creating $DIR/.vim/_temp..."
+  mkdir $DIR/.vim/_temp
+fi
+
+# Sourcing from .bash_profile
+echo "Source from ~/.bash_profile ..."
+source ~/.bash_profile
+
+# Done!
+echo "DONE!"
